// types/type.ts
export type CardDataType = {
    id: number;
    methodTH: string;
    methodEN: string;
    methodJP: string;
    titleTH: string;
    titleEN: string;
    titleJP: string;
    contentTH: string;
    contentEN: string;
    contentJP: string;
    timeRecommend: string;
    use: string;
    minTime: number;
    maxTime: number;
  };
  